backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pandas==2.2.2
scikit-learn==1.4.2
xgboost==2.0.3
shap==0.45.1
joblib==1.4.2
numpy==1.26.4
python-dotenv==1.0.1
httpx==0.27.0
groq==0.9.0
slowapi==0.1.9
slowapi==0.1.9
